Adjust experimental bans for non-pgm servers
