+++ src/main/java/dev/pgm/community/moderation/feature/PGMPunishmentIntegration.java
@@ -1,16 +1,30 @@
 package dev.pgm.community.moderation.feature;
 
+import dev.pgm.community.Community;
 import dev.pgm.community.moderation.punishments.Punishment;
 import dev.pgm.community.moderation.punishments.types.MutePunishment;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
+import tc.oc.pgm.api.event.NameDecorationChangeEvent;
+import tc.oc.pgm.api.integration.Integration;
 import tc.oc.pgm.api.integration.PunishmentIntegration;
 
 public class PGMPunishmentIntegration implements PunishmentIntegration {
 
+  private static final String BANNED_GROUP = "pgm.group.banned";
   private ModerationFeatureBase moderation;
 
   public PGMPunishmentIntegration(ModerationFeatureBase moderation) {
     this.moderation = moderation;
+  }
+
+  public void enable() {
+    Integration.setPunishmentIntegration(this);
+  }
+
+  public void updateBanPrefix(Player player, boolean apply) {
+    player.addAttachment(Community.get(), BANNED_GROUP, apply);
+    Bukkit.getPluginManager().callEvent(new NameDecorationChangeEvent(player.getUniqueId()));
   }
 
   @Override
