+++ src/main/java/dev/pgm/community/teleports/TeleportFeatureBase.java
@@ -14,9 +14,9 @@ import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Location;
 import org.bukkit.configuration.Configuration;
 import org.bukkit.entity.Player;
-import tc.oc.pgm.api.text.PlayerComponent;
 import tc.oc.pgm.util.Audience;
 import tc.oc.pgm.util.named.NameStyle;
+import tc.oc.pgm.util.text.PlayerComponent;
 
 public class TeleportFeatureBase extends FeatureBase implements TeleportFeature {
